frontend: Move apiProxy files to api/v1
