+++ frontend/src/lib/k8s/api/v1/factories.ts
@@ -2,9 +2,9 @@
 //       is 404 and not trying it again... and again.
 
 import { OpPatch } from 'json-patch';
-import { isDebugVerbose } from '../../../helpers';
-import { getCluster } from '../../cluster';
-import { KubeObjectInterface } from '../cluster';
+import { isDebugVerbose } from '../../../../helpers';
+import { getCluster } from '../../../cluster';
+import { KubeObjectInterface } from '../../cluster';
 import { ApiError, clusterRequest, patch, post, put, remove } from './clusterRequests';
 import { asQuery, getApiRoot } from './formatUrl';
 import { QueryParameters } from './queryParameters';
