+++ frontend/src/lib/k8s/api/v1/metricsApi.ts
@@ -1,6 +1,6 @@
-import { isDebugVerbose } from '../../../helpers';
-import { getCluster } from '../../cluster';
-import { KubeMetrics } from '../cluster';
+import { isDebugVerbose } from '../../../../helpers';
+import { getCluster } from '../../../cluster';
+import { KubeMetrics } from '../../cluster';
 import { ApiError, clusterRequest } from './clusterRequests';
 
 /**
