+++ frontend/src/lib/k8s/api/v1/streamingApi.ts
@@ -1,8 +1,8 @@
-import { isDebugVerbose } from '../../../helpers';
-import { findKubeconfigByClusterName, getUserIdFromLocalStorage } from '../../../stateless';
-import { getToken } from '../../auth';
-import { getCluster } from '../../cluster';
-import { KubeObjectInterface } from '../cluster';
+import { isDebugVerbose } from '../../../../helpers';
+import { findKubeconfigByClusterName, getUserIdFromLocalStorage } from '../../../../stateless';
+import { getToken } from '../../../auth';
+import { getCluster } from '../../../cluster';
+import { KubeObjectInterface } from '../../cluster';
 import { ApiError, clusterRequest } from './clusterRequests';
 import { BASE_HTTP_URL, CLUSTERS_PREFIX } from './constants';
 import { asQuery, combinePath } from './formatUrl';
